requirements-test.txt
# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.0
pytest-env>=1.0.0
requests>=2.31.0
boto3>=1.28.0
python-dotenv>=1.0.0

